Framework for App
Framework used for our 2D app - currently draws an icon to a screen and
lets it move in a random direction.  Also tells you the FPS of the
rendering.

The next phase would be to be able to move one icon through player
input.  Following that would be for the player to shoot and move via
player input.